refactor: remove `peer` module
